+++ R/socket.R
@@ -34,6 +34,9 @@
 #' @param autostart [default TRUE] whether to start the dialer/listener. Set to
 #'     FALSE if you wish to set configuration options on the dialer/listener as
 #'     it is not generally possible to change these once started.
+#' @param raw [default 'FALSE'] whether to open 'raw' mode sockets. Note: not for
+#'     general use - do not enable unless you have a specific need (refer to NNG
+#'     documentation).
 #'
 #' @return A Socket (object of class 'nanoSocket' and 'nano').
 #'
@@ -77,11 +80,12 @@ socket <- function(protocol = c("bus", "pair", "push", "pull", "pub", "sub",
                                 "req", "rep", "surveyor", "respondent"),
                    dial = NULL,
                    listen = NULL,
-                   autostart = TRUE) {
+                   autostart = TRUE,
+                   raw = FALSE) {
 
   protocol <- match.arg2(protocol, c("bus", "pair", "push", "pull", "pub", "sub",
                                      "req", "rep", "surveyor", "respondent"))
-  sock <- .Call(rnng_protocol_open, protocol)
+  sock <- .Call(rnng_protocol_open, protocol, raw)
   is.integer(sock) && return(sock)
   if (!missing(dial)) dial(sock, url = dial, autostart = autostart)
   if (!missing(listen)) listen(sock, url = listen, autostart = autostart)